corrigiendo errores del primer pull request de Angelo
