+++ app/.env.local
@@ -1,6 +1,8 @@
-NEXT_PUBLIC_FIREBASE_API_KEY=AIzaSyArFGqGZ4phKYOQsA4UJN4UotS2djQ8lUo
-NEXT_PUBLIC_FIREBASE_AUTH_DOMAIN=login-89a2d.firebaseapp.com
-NEXT_PUBLIC_FIREBASE_PROJECT_ID=login-89a2d
-NEXT_PUBLIC_FIREBASE_STORAGE_BUCKET=login-89a2d.appspot.com
-NEXT_PUBLIC_FIREBASE_MESSAGING_SENDER_ID=572463801506
-NEXT_PUBLIC_FIREBASE_APP_ID=1:572463801506:web:5335a8c1e7b4daf18ba3fb
+NEXT_PUBLIC_FIREBASE_PUBLIC_API_KEY=AIzaSyCXZ9qejpgjElD6sEIpH_zc3A_k2LKeDBQ
+NEXT_PUBLIC_FIREBASE_AUTH_DOMAIN=get-things-de986.firebaseapp.com
+NEXT_PUBLIC_FIREBASE_PROJECT_ID=get-things-de986
+
+# storageBucket: "get-things-de986.appspot.com",
+# messagingSenderId: "1070753380099",
+# appId: "1:1070753380099:web:74323baa109ca0aef95b36",
+# measurementId: "G-C1LEYLWKFC"
