+++ app/README.md
@@ -34,3 +34,24 @@ Font Awesome es la forma más popular de agregar íconos de fuentes a su sitio w
 
 Este proyecto está bajo la Licencia (MIT) - mira el archivo [LICENSE.md](LICENSE.md) para detalles
 
+## Geist-ui/React
+
+Introducción
+
+Acerca de
+
+Geist UI es una implementación de React para estilos procedente de Diseño de Vercel.
+
+El diseño de Vercel es conciso y estético, esta es una razón importante para la popularidad de Vercel. Ahora puede implementar los estilos en su propio proyecto React.
+
+Geist UI es un sistema de diseño de código abierto para crear sitios web y aplicaciones modernos, todos nuestros proyectos son de código abierto.
+
+## Firebase
+
+se verifico que cada login de red social funciona
+
+![github](https://user-images.githubusercontent.com/67810145/122855731-460f2300-d2db-11eb-9fe7-248b443924be.PNG)
+
+![google](https://user-images.githubusercontent.com/67810145/122856575-a2267700-d2dc-11eb-8c37-84285feec3ed.PNG)
+
+![facebook](https://user-images.githubusercontent.com/67810145/122856588-a8b4ee80-d2dc-11eb-8064-7e684ec24c67.PNG)
