+++ app/package.json
@@ -12,6 +12,7 @@
     "@fortawesome/free-brands-svg-icons": "5.15.3",
     "@fortawesome/free-solid-svg-icons": "5.15.3",
     "@fortawesome/react-fontawesome": "0.1.14",
+    "@geist-ui/react": "^2.1.5",
     "@types/node": "^15.12.3",
     "@types/react": "^17.0.11",
     "firebase": "^8.6.8",
